Fix gr_clip
gr_clip was not calculating the correct values after the rotation
patch. Since gr_clip width and height instead of taking a second
pair of x,y coordinates, we cannot depend as much on the
ROTATION_X_DISP and ROTATION_Y_DISP macros. The simplest solution
was to call scissor with direct values in most cases.

Change-Id: I25f9d575ec6ddc430fca4555f29c8334441fc95d